test_ad_risk is now passing with the new VCF and fixed the golddata
