+++ tests/test_ad_risk.py
@@ -1,11 +1,7 @@
-import os
 import unittest
 import sys
 from pathlib import Path
 import pandas as pd
-import numpy as np
-import stash as st
-import shutil
 # Add parent directory to path
 CURRENT_PATH = Path(__file__).parent
 sys.path.insert(0, str(CURRENT_PATH.parent))
@@ -17,6 +13,7 @@ _REPO_ROOT = Path(__file__).parent.parent.resolve()
 ARTIFACTS_DIR = _REPO_ROOT / '_artifacts'
 VCF_EXAMPLE = ARTIFACTS_DIR / "HG00096.vcf.gz"
 TISSUE_MAP_PATH = ARTIFACTS_DIR / "be73e19a.pq"
+APOE_GENE_ID = 'ENSG00000130203.9'
 
 
 class TestADrisk(unittest.TestCase):
@@ -25,10 +22,10 @@ class TestADrisk(unittest.TestCase):
         Set up the test case with a sample ADrisk instance and test data.
         """
         self.model_class = 'v4_pcg'
-        self.gene_id = 'ENSG00000130203.9'  # APOE gene
+        self.gene_id = APOE_GENE_ID
         self.tissue_id = 7
         self.adrisk = ad_risk.ADrisk(self.gene_id, self.tissue_id, model_class=self.model_class)
-        self.vcf_path = '/mnt/czi-sci-ai/intrinsic-variation-gene-ex-2/project_gene_regulation/dna2cell_training/v2_pcg_flash2/sample_vcf/HG00096.vcf.gz'
+        self.vcf_path = VCF_EXAMPLE
         self._init_vcf_processor()
         self._init_dataloader()
         self._init_vcf_processor()
@@ -50,15 +47,12 @@ class TestADrisk(unittest.TestCase):
 
     def test_1(self) -> None:
         """
-        Test the ADrisk prediction pipeline with sample data.
+        Test the ADrisk prediction from gene embedding
         """
-        preds_df = self.vcf_processor.predict(self.model, self.checkpoint_path, self.trainer, self.dataloader, self.vcf_dataset)
-        gene_tissue_embeds = preds_df['embeddings'].iloc[0]
-        # gene_tissue_embeds = np.vstack([gene_tissue_embeds, gene_tissue_embeds])
+        model_outputs = self.vcf_processor.predict(self.model, self.checkpoint_path, self.trainer, self.dataloader, self.vcf_dataset)
+        gene_tissue_embeds = model_outputs['embeddings'].iloc[0]
         preds = self.adrisk(gene_tissue_embeds)
-        import ipdb; ipdb.set_trace()
-        self.assertAlmostEqual(preds[0], 0.31081957, places=5) # <- Needs to be checked
-        print(f"AD risk proba prediction is: {preds}")
+        self.assertAlmostEqual(preds[0], 0.66763765, places=2) # tested on h100, dev coreweave on 10/31/25
 
 
 class TestADriskFromVCF(unittest.TestCase):
@@ -67,13 +61,13 @@ class TestADriskFromVCF(unittest.TestCase):
         Set up the test case with a sample ADrisk instance and test data.
         """
         self.adrisk = ad_risk.ADriskFromVCF()
-        self.vcf_path = 
-        self.gene_ids = ['ENSG00000000457.13'] * 2
-        self.tissue_ids = [43, 47]
+        self.vcf_path = VCF_EXAMPLE
+        self.gene_ids = [APOE_GENE_ID] * 2
+        self.tissue_ids = [7, 47] # run it on 2 tissues
 
     def test_1(self) -> None:
         """
-        Test the ADrisk prediction pipeline with sample data.
+        Test the ADrisk prediction pipeline with an input VCF file
         """
         preds = self.adrisk(self.vcf_path, self.gene_ids, self.tissue_ids)
-        print(preds[['gene_name', 'tissue_name', 'ad_risk']])
+        self.assertAlmostEqual(preds.iloc[0].ad_risk.item(), 0.66763765, places=2) # tested on h100, dev coreweave on 10/31/25
